requirements.txt
asgiref==3.11.1
black==26.1.0
click==8.3.1
Django==4.1
django-debug-toolbar==3.2.4
djangorestframework==3.13.1
flake8==5.0.4
flake8-quotes==3.3.1
flake8-variables-names==0.0.5
mccabe==0.7.0
mypy_extensions==1.1.0
packaging==26.0
pathspec==1.0.4
pep8-naming==0.13.2
pillow==12.1.0
platformdirs==4.5.1
pycodestyle==2.9.1
pyflakes==2.5.0
pytokens==0.4.1
pytz==2025.2
sqlparse==0.5.5
